netctl-base: Allow subproject connected as submodule to be installed
When installing for packaging there is good option to clone only project
connecting all other subprojects and install each of them separarely.

To make this work introduce new variable PARENT which is set to NAME of
project calling this subproject installation (i.e. parent project) or
empty in case we call installation directly.

Note that AS_SUBPROJECT variable is now obsolete and removed to avoid
confusion when installing subproject independently from project it is
connected as submodule.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>